requirements.txt
streamlit
apify-client
openai
python-dotenv
